g.download.project: rename g.download.location (#4187)
