+++ .github/workflows/test_thorough.sh
@@ -4,7 +4,7 @@
 set -e
 
 grass --tmp-project XY --exec \
-    g.download.location url=https://grass.osgeo.org/sampledata/north_carolina/nc_spm_full_v2alpha2.tar.gz path=$HOME
+    g.download.project url=https://grass.osgeo.org/sampledata/north_carolina/nc_spm_full_v2alpha2.tar.gz path=$HOME
 
 grass --tmp-project XY --exec \
     python3 -m grass.gunittest.main \
