+++ scripts/Makefile
@@ -19,6 +19,7 @@ SUBDIRS = \
 	db.test \
 	db.univar \
 	g.download.location \
+	g.download.project \
 	g.extension \
 	g.extension.all \
 	g.manual \
