+++ scripts/g.download.project/Makefile
@@ -1,0 +1,7 @@
+MODULE_TOPDIR = ../..
+
+PGM = g.download.project
+
+include $(MODULE_TOPDIR)/include/Make/Script.make
+
+default: script
